server/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
twilio==8.10.0
google-cloud-firestore==2.16.0
google-auth==2.34.0
pytest-cov==5.0.0
redis==5.0.7
aioredis==2.0.1
python-jose[cryptography]==3.3.0
